backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt
pyjwt
